+++ utils/utils_windows.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"os"
 	"os/exec"
+	"regexp"
 	"strconv"
 	"strings"
 )
@@ -25,17 +26,23 @@ func DirWritable(path string) (bool, error) {
 	return true, nil
 }
 
-func Ping(address string, secondsTimeout int) error {
+func Ping(address string, secondsTimeout int) (int64, error) {
 	ping, err := exec.LookPath("ping")
 	if err != nil {
-		return err
+		return 0, err
 	}
 	out, _, err := Command(ping, address, "-n", "1", "-w", strconv.Itoa(secondsTimeout*1000))
 	if err != nil {
-		return err
+		return 0, err
 	}
 	if strings.Contains(out, "Destination Host Unreachable") {
-		return errors.New("destination host unreachable")
+		return 0, errors.New("destination host unreachable")
 	}
-	return nil
+	r := regexp.MustCompile(`Average = (.*)ms`)
+	strs := r.FindStringSubmatch(out)
+	if len(strs) < 2 {
+		return 0, errors.New("could not parse ping duration")
+	}
+	f, _ := strconv.ParseFloat(strs[1], 64)
+	return int64(f * 1000), nil
 }